Package refactoring for beans
