+++ textdb/textdb-web/src/main/java/edu/uci/ics/textdb/web/request/beans/DictionaryMatcherBean.java
@@ -1,10 +1,8 @@
-package edu.uci.ics.textdb.web.request.operatorbean;
+package edu.uci.ics.textdb.web.request.beans;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import edu.uci.ics.textdb.common.constants.DataConstants;
 import edu.uci.ics.textdb.plangen.operatorbuilder.DictionaryMatcherBuilder;
-import edu.uci.ics.textdb.plangen.operatorbuilder.OperatorBuilderUtils;
-import edu.uci.ics.textdb.web.request.OperatorBean;
 
 import java.util.HashMap;
