+++ textdb/textdb-web/src/main/java/edu/uci/ics/textdb/web/request/beans/FuzzyTokenSourceBean.java
@@ -1,29 +1,33 @@
-package edu.uci.ics.textdb.web.request.operatorbean;
+package edu.uci.ics.textdb.web.request.beans;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import edu.uci.ics.textdb.plangen.operatorbuilder.FuzzyTokenMatcherBuilder;
-import edu.uci.ics.textdb.web.request.OperatorBean;
+import edu.uci.ics.textdb.plangen.operatorbuilder.OperatorBuilderUtils;
 
 import java.util.HashMap;
 
 /**
- * This class defines the properties/data members specific to the FuzzyTokenMatcher operator
+ * This class defines the properties/data members specific to the FuzzyTokenSource operator
  * and extends the OperatorBean class which defines the data members general to all operators
  * Created by kishorenarendran on 10/17/16.
  */
-public class FuzzyTokenMatcherBean extends OperatorBean {
+public class FuzzyTokenSourceBean extends OperatorBean {
+
     @JsonProperty("query")
     private String query;
     @JsonProperty("threshold_ratio")
     private String thresholdRatio;
+    @JsonProperty("data_source")
+    private String dataSource;
 
-    public FuzzyTokenMatcherBean() {
+    public FuzzyTokenSourceBean() {
     }
 
-    public FuzzyTokenMatcherBean(String operatorID, String operatorType, String query, String thresholdRatio) {
+    public FuzzyTokenSourceBean(String operatorID, String operatorType, String query, String thresholdRatio, String dataSource) {
         super(operatorID, operatorType);
         this.query = query;
         this.thresholdRatio = thresholdRatio;
+        this.dataSource = dataSource;
     }
 
     @JsonProperty("query")
@@ -46,10 +50,21 @@ public class FuzzyTokenMatcherBean extends OperatorBean {
         this.thresholdRatio = thresholdRatio;
     }
 
+    @JsonProperty("data_source")
+    public String getDataSource() {
+        return dataSource;
+    }
+
+    @JsonProperty("data_source")
+    public void setDataSource(String dataSource) {
+        this.dataSource = dataSource;
+    }
+
     public HashMap<String, String> getOperatorProperties() {
         HashMap<String, String> operatorProperties = super.getOperatorProperties();
         operatorProperties.put(FuzzyTokenMatcherBuilder.FUZZY_STRING, this.getQuery());
         operatorProperties.put(FuzzyTokenMatcherBuilder.THRESHOLD_RATIO, this.getThresholdRatio());
+        operatorProperties.put(OperatorBuilderUtils.DATA_DIRECTORY, this.getDataSource());
         return operatorProperties;
     }
 }
