+++ textdb/textdb-web/src/main/java/edu/uci/ics/textdb/web/request/beans/KeywordMatcherBean.java
@@ -1,9 +1,8 @@
-package edu.uci.ics.textdb.web.request.operatorbean;
+package edu.uci.ics.textdb.web.request.beans;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import edu.uci.ics.textdb.common.constants.DataConstants.KeywordMatchingType;
 import edu.uci.ics.textdb.plangen.operatorbuilder.KeywordMatcherBuilder;
-import edu.uci.ics.textdb.web.request.OperatorBean;
 
 import java.util.HashMap;
