+++ textdb/textdb-web/src/main/java/edu/uci/ics/textdb/web/request/beans/KeywordSourceBean.java
@@ -1,45 +1,43 @@
-package edu.uci.ics.textdb.web.request.operatorbean;
+package edu.uci.ics.textdb.web.request.beans;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import edu.uci.ics.textdb.common.constants.DataConstants;
-import edu.uci.ics.textdb.plangen.operatorbuilder.DictionaryMatcherBuilder;
-import edu.uci.ics.textdb.plangen.operatorbuilder.DictionarySourceBuilder;
+import edu.uci.ics.textdb.plangen.operatorbuilder.KeywordMatcherBuilder;
 import edu.uci.ics.textdb.plangen.operatorbuilder.OperatorBuilderUtils;
-import edu.uci.ics.textdb.web.request.OperatorBean;
 
 import java.util.HashMap;
 
 /**
- * This class defines the properties/data members specific to the DictionarySource operator
+ * This class defines the properties/data members specific to the KeywordSource operator
  * and extends the OperatorBean class which defines the data members general to all operators
  * Created by kishorenarendran on 10/17/16.
  */
-public class DictionarySourceBean extends OperatorBean {
-    @JsonProperty("dictionary")
-    private String dictionary;
+public class KeywordSourceBean extends OperatorBean {
+    @JsonProperty("keyword")
+    private String keyword;
     @JsonProperty("matching_type")
     private DataConstants.KeywordMatchingType matchingType;
     @JsonProperty("data_source")
     private String dataSource;
 
-    public DictionarySourceBean() {
+    public KeywordSourceBean() {
     }
 
-    public DictionarySourceBean(String operatorID, String operatorType, String dictionary, DataConstants.KeywordMatchingType matchingType, String dataSource) {
+    public KeywordSourceBean(String operatorID, String operatorType, String keyword, DataConstants.KeywordMatchingType matchingType, String dataSource) {
         super(operatorID, operatorType);
-        this.dictionary = dictionary;
+        this.keyword = keyword;
         this.matchingType = matchingType;
         this.dataSource = dataSource;
     }
 
-    @JsonProperty("dictionary")
-    public String getDictionary() {
-        return dictionary;
+    @JsonProperty("keyword")
+    public String getKeyword() {
+        return keyword;
     }
 
-    @JsonProperty("dictionary")
-    public void setDictionary(String dictionary) {
-        this.dictionary = dictionary;
+    @JsonProperty("keyword")
+    public void setKeyword(String keyword) {
+        this.keyword = keyword;
     }
 
     @JsonProperty("matching_type")
@@ -64,8 +62,8 @@ public class DictionarySourceBean extends OperatorBean {
 
     public HashMap<String, String> getOperatorProperties() {
         HashMap<String, String> operatorProperties = super.getOperatorProperties();
-        operatorProperties.put(DictionarySourceBuilder.DICTIONARY, this.getDictionary());
-        operatorProperties.put(DictionarySourceBuilder.MATCHING_TYPE, this.getMatchingType().name());
+        operatorProperties.put(KeywordMatcherBuilder.KEYWORD, this.getKeyword());
+        operatorProperties.put(KeywordMatcherBuilder.MATCHING_TYPE, this.getMatchingType().name());
         operatorProperties.put(OperatorBuilderUtils.DATA_DIRECTORY, this.getDataSource());
         return operatorProperties;
     }
