+++ textdb/textdb-web/src/main/java/edu/uci/ics/textdb/web/request/beans/RegexSourceBean.java
@@ -1,9 +1,8 @@
-package edu.uci.ics.textdb.web.request.operatorbean;
+package edu.uci.ics.textdb.web.request.beans;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import edu.uci.ics.textdb.plangen.operatorbuilder.OperatorBuilderUtils;
 import edu.uci.ics.textdb.plangen.operatorbuilder.RegexMatcherBuilder;
-import edu.uci.ics.textdb.web.request.OperatorBean;
 
 import java.util.HashMap;
